Extract pidfile management into lib/cluster/pidfile.js
