+++ lib/cluster/index.js
@@ -3,11 +3,10 @@ ComboHandler.Cluster.Master
 **/
 var cluster = require('cluster');
 var util = require('util');
-var fs = require('fs');
-var path = require('path');
 
 var ComboBase = require('./base');
 var checkStatus = require('./status');
+var pidfiles = require('./pidfiles');
 
 exports = module.exports = ComboCluster;
 
@@ -89,14 +88,7 @@ ComboCluster.prototype._detachEvents = function () {
 };
 
 ComboCluster.prototype._setupMasterPidFile = function () {
-    var dir = this.options.pids;
-
-    // ensure pids dir exists before writing the pidfile
-    if (!fs.existsSync(dir)) {
-        require('mkdirp').sync(dir);
-    }
-
-    writePidFile(dir, 'master', process.pid);
+    pidfiles.writePidFile(this.options.pids, 'master', process.pid);
 };
 
 ComboCluster.prototype._bindProcess = function () {
@@ -131,7 +123,7 @@ ComboCluster.prototype._bindCluster = function () {
 
         // this doesn't work in OS X, but whatever
         worker.process.title = 'combohandler worker';
-        writePidFile(pids, 'worker' + worker.id, worker.process.pid);
+        pidfiles.writePidFile(pids, 'worker' + worker.id, worker.process.pid);
     });
 
     cluster.on('disconnect', function (worker) {
@@ -148,14 +140,14 @@ ComboCluster.prototype._bindCluster = function () {
 
         if (worker.suicide) {
             console.error('Worker %d exited cleanly.', worker.id);
-            removePidFile(pids, 'worker' + worker.id);
+            pidfiles.removePidFile(pids, 'worker' + worker.id);
         }
         else {
             if (signal) {
                 console.error('Worker %d received signal %s', worker.id, signal);
                 if (signal === 'SIGUSR2') {
                     console.error('Worker %d restarting, removing old pidfile', worker.id);
-                    removePidFile(pids, 'worker' + worker.id);
+                    pidfiles.removePidFile(pids, 'worker' + worker.id);
                 }
             }
 
@@ -163,7 +155,7 @@ ComboCluster.prototype._bindCluster = function () {
                 console.error('Worker %d exited with code %d', worker.id, code);
                 if (++flameouts > 20) {
                     console.error("Too many errors during startup, bailing!");
-                    removePidFileSync(pids, 'master');
+                    pidfiles.removePidFileSync(pids, 'master');
                     process.exit(1);
                 }
             }
@@ -181,7 +173,7 @@ ComboCluster.prototype.gracefulShutdown = function () {
     var dir = this.options.pids;
     cluster.disconnect(function () {
         process.on('exit', function () {
-            removePidFileSync(dir, 'master');
+            pidfiles.removePidFileSync(dir, 'master');
             console.log('combohandler master %d finished shutting down!', process.pid);
         });
     });
@@ -198,8 +190,7 @@ ComboCluster.prototype.restartWorkers = function () {
 };
 
 ComboCluster.prototype._signalMaster = function (signal) {
-    var dir = this.options.pids;
-    getMasterPid(dir, function (err, masterPid) {
+    pidfiles.getMasterPid(this.options.pids, function (err, masterPid, dir) {
         if (err) {
             console.error("Error sending signal %s to combohandler master process", signal);
             if ('ENOENT' === err.code) {
@@ -214,7 +205,7 @@ ComboCluster.prototype._signalMaster = function (signal) {
                 // again, because SIGKILL is so incredibly rude,
                 // he doesn't allow us to do anything afterward
                 if ('SIGKILL' === signal) {
-                    removePidFileSync(dir, 'master');
+                    pidfiles.removePidFileSync(dir, 'master');
                 }
                 // send signal to master process, not necessarily "killing" it
                 process.kill(masterPid, signal);
@@ -236,8 +227,7 @@ ComboCluster.prototype.restart = function () {
 };
 
 ComboCluster.prototype.status = function () {
-    var dir = this.options.pids;
-    getMasterPid(dir, function (err, masterPid) {
+    pidfiles.getMasterPid(this.options.pids, function (err, masterPid, dir) {
         if (err) {
             if ('ENOENT' === err.code) {
                 console.error('combohandler master not running!');
@@ -248,7 +238,7 @@ ComboCluster.prototype.status = function () {
             }
         }
 
-        getWorkerPidsSync(dir).forEach(checkStatus.bind({}, 'worker'));
+        pidfiles.getWorkerPidsSync(dir).forEach(checkStatus.bind(null, 'worker'));
 
         // check master last so we can exit non-zero when dead
         checkStatus('master', masterPid, '');
@@ -260,9 +250,9 @@ ComboCluster.prototype.shutdown = function () {
 };
 
 ComboCluster.prototype.stop = function () {
-    // must clean up worker pidfiles before sending SIGKILL,
+    // must clean up worker pid files before sending SIGKILL,
     // because SIGKILL listeners basically can't do anything
-    removeWorkerPidFiles(this.options.pids, this._stop.bind(this));
+    pidfiles.removeWorkerPidFiles(this.options.pids, this._stop.bind(this));
 };
 
 ComboCluster.prototype._stop = function () {
@@ -281,86 +271,3 @@ ComboCluster.prototype._listen = function () {
         cluster.fork();
     }
 };
-
-// Utilities ----------------------------------------------------------------
-
-// Manage combohandler process master and worker pidfiles
-// https://github.com/LearnBoost/cluster/blob/master/lib/plugins/pidfiles.js
-function getPidFilePath(dir, name) {
-    return path.join(dir, (name || 'master') + '.pid');
-}
-
-function getMasterPid(dir, cb) {
-    fs.readFile(getPidFilePath(dir, 'master'), function (err, pid) {
-        if (pid) {
-            pid = parseInt(pid, 10);
-        }
-        if (cb) {
-            cb(err, pid);
-        }
-    });
-}
-
-function getWorkerPidFilesSync(dir) {
-    return fs.readdirSync(dir).filter(function (file) {
-        return file.match(/^worker.*\.pid$/);
-    });
-}
-
-function getWorkerPidsSync(dir) {
-    return getWorkerPidFilesSync(dir).map(function (file) {
-        return parseInt(fs.readFileSync(dir + '/' + file), 10);
-    });
-}
-
-function writePidFile(dir, name, pid) {
-    fs.writeFile(getPidFilePath(dir, name), pid.toString(), function (err) {
-        if (err) { throw err; }
-    });
-}
-
-function removePidFile(dir, name, cb) {
-    fs.unlink(getPidFilePath(dir, name), function (err) {
-        if (cb) {
-            cb(err);
-        }
-        else if (err) {
-            if ('ENOENT' === err.code) {
-                console.error('Could not find pidfile: %s', err.msg);
-            }
-            else {
-                throw err;
-            }
-        }
-    });
-}
-
-function removePidFileSync(dir, name) {
-    fs.unlinkSync(getPidFilePath(dir, name));
-}
-
-function removeWorkerPidFiles(dir, cb) {
-    var workerPidFiles = getWorkerPidFilesSync(dir),
-        remaining = workerPidFiles.length;
-
-    if (!remaining && cb) {
-        cb();
-    }
-
-    workerPidFiles.forEach(function (file) {
-        removePidFile(dir, file.replace(/\.pid$/, ''), function (err) {
-            if (err) {
-                if ('ENOENT' === err.code) {
-                    console.error('Could not find worker pidfile: %s', file);
-                }
-                else {
-                    throw err;
-                }
-            }
-            if (--remaining === 0 && cb) {
-                cb();
-            }
-        });
-    });
-}
-
